# 📦 Latest Package Versions (June 12, 2025) & lxml Installation Fix