--- a/LXML_INSTALLATION_GUIDE.md
+++ b/LXML_INSTALLATION_GUIDE.md
@@ -1,0 +1,196 @@
+# lxml Installation Guide for Windows
+
+## Problem Description
+You're encountering an lxml installation error on Windows with Python 3.13. This is a common issue because lxml requires libxml2 and libxslt libraries to be compiled, which can be problematic on Windows.
+
+## Error Analysis
+The error shows:
+```
+fatal error C1083: Cannot open include file: 'libxml/xmlversion.h': No such file or directory
+```
+
+This indicates that the required libxml2 development headers are not available during compilation.
+
+## Solutions (in order of recommendation)
+
+### Solution 1: Use Pre-compiled Wheel (Recommended)
+```bash
+# Download the appropriate wheel for your Python version from:
+# https://pypi.org/project/lxml/#files
+
+# For Python 3.13 on Windows 64-bit:
+pip install https://files.pythonhosted.org/packages/.../lxml-5.4.0-cp313-cp313-win_amd64.whl
+
+# Or try installing directly (pip will find the wheel):
+pip install lxml==5.4.0 --only-binary=lxml
+```
+
+### Solution 2: Use Alternative XML Libraries
+Instead of lxml, use these alternatives that are easier to install on Windows:
+
+```bash
+# Install alternative XML processing libraries
+pip install html5lib==1.1
+pip install defusedxml==0.7.1
+pip install xmltodict==0.13.0
+pip install untangle==1.2.1
+```
+
+### Solution 3: Install Build Dependencies (Advanced)
+If you need lxml specifically:
+
+```bash
+# Install Microsoft C++ Build Tools
+# Download from: https://visualstudio.microsoft.com/visual-cpp-build-tools/
+
+# Install libxml2 and libxslt using conda
+conda install lxml
+
+# Or use the Windows Subsystem for Linux (WSL)
+```
+
+### Solution 4: Use Our Windows-Compatible Requirements
+```bash
+# Use our pre-configured Windows requirements file
+pip install -r requirements-windows-2025.txt
+```
+
+## Code Modifications for lxml Alternatives
+
+If you're using lxml in your code, here are the alternatives:
+
+### Original lxml code:
+```python
+from lxml import etree, html
+```
+
+### Alternative using html5lib:
+```python
+import html5lib
+from xml.etree import ElementTree as ET
+
+# For HTML parsing
+def parse_html(content):
+    return html5lib.parse(content, treebuilder="etree")
+
+# For XML parsing
+def parse_xml(content):
+    return ET.fromstring(content)
+```
+
+### Alternative using defusedxml:
+```python
+import defusedxml.ElementTree as ET
+
+# Safe XML parsing
+def parse_xml_safe(content):
+    return ET.fromstring(content)
+```
+
+## Testing Your Installation
+
+Create a test file `test_xml.py`:
+
+```python
+#!/usr/bin/env python3
+"""Test XML processing capabilities"""
+
+def test_xml_libraries():
+    """Test available XML libraries"""
+    results = {}
+    
+    # Test html5lib
+    try:
+        import html5lib
+        results['html5lib'] = "✅ Available"
+    except ImportError:
+        results['html5lib'] = "❌ Not available"
+    
+    # Test defusedxml
+    try:
+        import defusedxml.ElementTree
+        results['defusedxml'] = "✅ Available"
+    except ImportError:
+        results['defusedxml'] = "❌ Not available"
+    
+    # Test lxml
+    try:
+        import lxml.etree
+        results['lxml'] = "✅ Available"
+    except ImportError:
+        results['lxml'] = "❌ Not available"
+    
+    # Test built-in xml
+    try:
+        import xml.etree.ElementTree
+        results['xml.etree'] = "✅ Available (built-in)"
+    except ImportError:
+        results['xml.etree'] = "❌ Not available"
+    
+    print("XML Library Status:")
+    for lib, status in results.items():
+        print(f"  {lib}: {status}")
+    
+    return results
+
+if __name__ == "__main__":
+    test_xml_libraries()
+```
+
+Run the test:
+```bash
+python test_xml.py
+```
+
+## Recommended Approach for Custom Gemini Agent GUI
+
+For the Custom Gemini Agent GUI project, we recommend:
+
+1. **Use the Windows-compatible requirements file**:
+   ```bash
+   pip install -r requirements-windows-2025.txt
+   ```
+
+2. **Modify the code to use alternative XML libraries** where lxml was used
+
+3. **Update the web scraping components** to use html5lib instead of lxml
+
+## Updated Installation Commands
+
+```bash
+# Create virtual environment
+python -m venv venv
+venv\Scripts\activate
+
+# Upgrade pip and install build tools
+python -m pip install --upgrade pip setuptools wheel
+
+# Install Windows-compatible requirements
+pip install -r requirements-windows-2025.txt
+
+# Verify installation
+python verify_installation.py
+```
+
+## If You Still Need lxml
+
+If lxml is absolutely required, try these steps:
+
+1. **Install Visual Studio Build Tools**:
+   - Download from: https://visualstudio.microsoft.com/visual-cpp-build-tools/
+   - Install "C++ build tools" workload
+
+2. **Use conda instead of pip**:
+   ```bash
+   conda install -c conda-forge lxml
+   ```
+
+3. **Use pre-compiled binaries**:
+   ```bash
+   # Download from https://www.lfd.uci.edu/~gohlke/pythonlibs/#lxml
+   pip install lxml-5.4.0-cp313-cp313-win_amd64.whl
+   ```
+
+## Summary
+
+The lxml installation issue is common on Windows. Our recommended solution is to use alternative XML libraries that are easier to install and maintain. The Custom Gemini Agent GUI has been designed to work with these alternatives for better Windows compatibility.