--- a/requirements-windows-2025.txt
+++ b/requirements-windows-2025.txt
@@ -1,0 +1,85 @@
+# Custom Gemini Agent GUI - Windows Compatible Requirements (June 12, 2025)
+# Optimized for Windows installation with Python 3.13 compatibility
+
+# Core GUI Framework
+PyQt6==6.9.1
+
+# AI and ML Libraries
+sentence-transformers==4.1.0
+google-generativeai==0.8.3
+
+# Vector Database
+chromadb==1.0.12
+
+# Document Processing and RAG
+langchain==0.3.7
+langchain-community==0.3.7
+langchain-core==0.3.7
+
+# File System Monitoring
+watchdog==5.0.3
+
+# Git Integration
+GitPython==3.1.43
+
+# Google Drive API
+google-api-python-client==2.145.0
+google-auth-httplib2==0.2.0
+google-auth-oauthlib==1.2.1
+
+# Document Loaders
+PyPDF2==3.0.1
+python-docx==1.1.2
+openpyxl==3.1.5
+
+# Web Scraping and HTTP
+requests==2.32.3
+beautifulsoup4==4.12.3
+trafilatura==1.12.2
+
+# XML Processing - Windows Compatible Alternative
+# Instead of lxml which has compilation issues on Windows
+html5lib==1.1
+defusedxml==0.7.1
+
+# Secure Storage
+keyring==25.4.1
+
+# Configuration Management
+pydantic==2.11.5
+pydantic-settings==2.6.1
+
+# Logging
+loguru==0.7.2
+
+# Testing
+pytest==8.3.3
+pytest-qt==4.4.0
+pytest-mock==3.14.0
+
+# Development Tools
+black==24.10.0
+flake8==7.1.1
+mypy==1.13.0
+
+# Packaging
+PyInstaller==6.14.1
+
+# Windows-specific dependencies
+pywin32==306
+pywin32-ctypes==0.2.3
+
+# Core Python Dependencies
+typing-extensions==4.12.2
+setuptools==75.1.0
+wheel==0.44.0
+packaging==24.1
+six==1.16.0
+
+# Additional Windows compatibility
+colorama==0.4.6
+wincertstore==0.2.1
+
+# Alternative XML libraries for Windows
+xmltodict==0.13.0
+untangle==1.2.1